refactor(web): register.d.ts => types.ts
JIP-57
